+++ familyhelper-note-impl/src/main/java/com/dwarfeng/familyhelper/note/impl/dao/preset/NoteItemPresetCriteriaMaker.java
@@ -30,11 +30,20 @@ public class NoteItemPresetCriteriaMaker implements PresetCriteriaMaker {
             case NoteItemMaintainService.NAME_LIKE:
                 nameLike(detachedCriteria, objects);
                 break;
+            case NoteItemMaintainService.CHILD_FOR_BOOK_NAME_LIKE:
+                childForBookNameLike(detachedCriteria, objects);
+                break;
             case NoteItemMaintainService.CHILD_FOR_NODE_INDEX_ASC:
                 childForNodeIndexAsc(detachedCriteria, objects);
                 break;
             case NoteItemMaintainService.CHILD_FOR_NODE_INDEX_DESC:
                 childForNodeIndexDesc(detachedCriteria, objects);
+                break;
+            case NoteItemMaintainService.CHILD_FOR_BOOK_ROOT_INDEX_ASC:
+                childForBookRootIndexAsc(detachedCriteria, objects);
+                break;
+            case NoteItemMaintainService.CHILD_FOR_BOOK_ROOT_INDEX_DESC:
+                childForBookRootIndexDesc(detachedCriteria, objects);
                 break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + s);
@@ -97,6 +106,24 @@ public class NoteItemPresetCriteriaMaker implements PresetCriteriaMaker {
         }
     }
 
+    @SuppressWarnings("DuplicatedCode")
+    private void childForBookNameLike(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isNull("bookLongId"));
+            } else {
+                LongIdKey longIdKey = (LongIdKey) objects[0];
+                detachedCriteria.add(
+                        Restrictions.eqOrIsNull("bookLongId", longIdKey.getLongId())
+                );
+            }
+            String pattern = (String) objects[1];
+            detachedCriteria.add(Restrictions.like("name", pattern, MatchMode.ANYWHERE));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
     private void childForNodeIndexAsc(DetachedCriteria detachedCriteria, Object[] objects) {
         try {
             if (Objects.isNull(objects[0])) {
@@ -123,6 +150,40 @@ public class NoteItemPresetCriteriaMaker implements PresetCriteriaMaker {
                         Restrictions.eqOrIsNull("nodeLongId", longIdKey.getLongId())
                 );
             }
+            detachedCriteria.addOrder(Order.desc("index"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void childForBookRootIndexAsc(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isNull("bookLongId"));
+            } else {
+                LongIdKey longIdKey = (LongIdKey) objects[0];
+                detachedCriteria.add(
+                        Restrictions.eqOrIsNull("bookLongId", longIdKey.getLongId())
+                );
+            }
+            detachedCriteria.add(Restrictions.isNull("nodeLongId"));
+            detachedCriteria.addOrder(Order.asc("index"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void childForBookRootIndexDesc(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isNull("bookLongId"));
+            } else {
+                LongIdKey longIdKey = (LongIdKey) objects[0];
+                detachedCriteria.add(
+                        Restrictions.eqOrIsNull("bookLongId", longIdKey.getLongId())
+                );
+            }
+            detachedCriteria.add(Restrictions.isNull("nodeLongId"));
             detachedCriteria.addOrder(Order.desc("index"));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));