+++ familyhelper-note-impl/src/main/java/com/dwarfeng/familyhelper/note/impl/dao/preset/NoteNodePresetCriteriaMaker.java
@@ -4,6 +4,7 @@ import com.dwarfeng.familyhelper.note.stack.service.NoteNodeMaintainService;
 import com.dwarfeng.subgrade.sdk.hibernate.criteria.PresetCriteriaMaker;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.MatchMode;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Component;
 
@@ -24,6 +25,12 @@ public class NoteNodePresetCriteriaMaker implements PresetCriteriaMaker {
                 break;
             case NoteNodeMaintainService.CHILD_FOR_BOOK_ROOT:
                 childForBookRoot(detachedCriteria, objects);
+                break;
+            case NoteNodeMaintainService.NAME_LIKE:
+                nameLike(detachedCriteria, objects);
+                break;
+            case NoteNodeMaintainService.CHILD_FOR_BOOK_NAME_LIKE:
+                childForBookNameLike(detachedCriteria, objects);
                 break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + s);
@@ -72,6 +79,33 @@ public class NoteNodePresetCriteriaMaker implements PresetCriteriaMaker {
                 );
             }
             detachedCriteria.add(Restrictions.isNull("parentLongId"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void nameLike(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            String pattern = (String) objects[0];
+            detachedCriteria.add(Restrictions.like("name", pattern, MatchMode.ANYWHERE));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    private void childForBookNameLike(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            if (Objects.isNull(objects[0])) {
+                detachedCriteria.add(Restrictions.isNull("bookLongId"));
+            } else {
+                LongIdKey longIdKey = (LongIdKey) objects[0];
+                detachedCriteria.add(
+                        Restrictions.eqOrIsNull("bookLongId", longIdKey.getLongId())
+                );
+            }
+            String pattern = (String) objects[1];
+            detachedCriteria.add(Restrictions.like("name", pattern, MatchMode.ANYWHERE));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }